+++ djrill/urls.py
@@ -1,13 +1,11 @@
 try:
-    from django.conf.urls import patterns, url
+    from django.conf.urls import url
 except ImportError:
-    from django.conf.urls.defaults import patterns, url
+    from django.conf.urls.defaults import url
 
 from .views import DjrillWebhookView
 
 
-urlpatterns = patterns(
-    '',
-
+urlpatterns = [
     url(r'^webhook/$', DjrillWebhookView.as_view(), name='djrill_webhook'),
-)
+]